main/fav/hist: Add support for nameable favourites
Old data migration:
For untitled favourites, we show "<Untitled>" to the
user.

For new data:
If it is from history, we show "<Untitled>" to the
user as well, as we don't save the title of a page
in history right now.
If it is saved from Main, it will include the
webpage's title.

Signed-off-by: Ricky Cheung <rcheung844@gmail.com>