main: java: Save beforeNextUrl to url
The purpose for this is to save the Url before the error
screen shows up, so saving a proper Url should be done.

This also fixes a null reference, it was saving null as
the website because the code that loads the default website
on start also uses this function, if it tries to save it,
webview will return null because nothing is loaded.

Signed-off-by: Ricky Cheung <rcheung844@gmail.com>