RESTRICT AUTOMERGE viola: search: Add FrogFind!
How can we not have "The Search Engine for Vintage Computers" on
the list? So that people can leap without JavaScript!

Need to confirm whether suggestions is disabled after this. Tests
required to port downstream. Also, we need to add a proper way
to convert this back for those updating. Consider using IDs this
time around instead of depending on index. This is not the 1960s!

Change-Id: I0c7616ab952f6e401887e5ebe7b5f33ded49e483